+++ amfiteatr_core/src/error/convert.rs
@@ -3,5 +3,5 @@ use thiserror::Error;
 #[derive(Debug, Clone, Error)]
 pub enum ConvertError{
     #[error("Action convert from {0}")]
-    ActionDeserialize(String)
+    ConvertFromTensor(String)
 }