+++ amfiteatr_examples/examples/cart_pole_python/common.rs
@@ -75,13 +75,13 @@ impl TryFrom<&Tensor> for CartPoleAction{
         let v: Vec<i64> = match Vec::try_from(tensor){
             Ok(v) => v,
             Err(_) =>{
-                return Err(ConvertError::ActionDeserialize(format!("{}", tensor)))
+                return Err(ConvertError::ConvertFromTensor(format!("{}", tensor)))
             }
         };
         match v[0]{
             0 => Ok(CartPoleAction::Left),
             1 => Ok(CartPoleAction::Right),
-            _ => Err(ConvertError::ActionDeserialize(format!("{}", tensor)))
+            _ => Err(ConvertError::ConvertFromTensor(format!("{}", tensor)))
         }
     }
 }