+++ amfiteatr_rl/src/policy/mod.rs
@@ -3,6 +3,7 @@ mod q_learning_policy;
 //mod experiencing_policy;
 mod learning_policy;
 mod train_config;
+mod ppo;
 //mod network_layer;
 
 pub use actor_critic::*;
@@ -10,3 +11,4 @@ pub use q_learning_policy::*;
 //pub use experiencing_policy::*;
 pub use learning_policy::*;
 pub use train_config::*;
+pub use ppo::*;