+++ amfiteatr_rl/src/policy/ppo/mod.rs
@@ -1,0 +1,2 @@
+mod multi_discrete;
+pub use multi_discrete::*;